@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": "^7.4||^8.0",
+        "php": "^8.0",
         "bacon/bacon-qr-code": "^2.0.5"
     },
     "require-dev": {
@@ -41,6 +41,9 @@
         "sort-packages": true,
         "preferred-install": {
             "endroid/*": "source"
+        },
+        "allow-plugins": {
+            "endroid/installer": true
         }
     },
     "extra": {
