@@ -8,19 +8,14 @@ use Endroid\QrCode\Matrix\MatrixInterface;
 
 final class PngResult extends AbstractResult
 {
-    /** @var mixed */
-    private $image;
-
-    /** @param mixed $image */
-    public function __construct(MatrixInterface $matrix, $image)
-    {
+    public function __construct(
+        MatrixInterface $matrix,
+        private \GdImage $image
+    ) {
         parent::__construct($matrix);
-
-        $this->image = $image;
     }
 
-    /** @return mixed */
-    public function getImage()
+    public function getImage(): \GdImage
     {
         return $this->image;
     }
