@@ -6,15 +6,12 @@ namespace Endroid\QrCode\Encoding;
 
 final class Encoding implements EncodingInterface
 {
-    private string $value;
-
-    public function __construct(string $value)
-    {
+    public function __construct(
+        private string $value
+    ) {
         if (!in_array($value, mb_list_encodings())) {
             throw new \Exception(sprintf('Invalid encoding "%s"', $value));
         }
-
-        $this->value = $value;
     }
 
     public function __toString(): string
