@@ -8,13 +8,10 @@ use Endroid\QrCode\Label\LabelInterface;
 
 class LabelImageData
 {
-    private int $width;
-    private int $height;
-
-    private function __construct(int $width, int $height)
-    {
-        $this->width = $width;
-        $this->height = $height;
+    private function __construct(
+        private int $width,
+        private int $height
+    ) {
     }
 
     public static function createForLabel(LabelInterface $label): self
