@@ -8,9 +8,9 @@ interface LogoInterface
 {
     public function getPath(): string;
 
-    public function getResizeToWidth(): ?int;
+    public function getResizeToWidth(): int|null;
 
-    public function getResizeToHeight(): ?int;
+    public function getResizeToHeight(): int|null;
 
     public function getPunchoutBackground(): bool;
 }
