@@ -15,30 +15,24 @@ use Endroid\QrCode\RoundBlockSizeMode\RoundBlockSizeModeMargin;
 
 final class QrCode implements QrCodeInterface
 {
-    private string $data;
     private EncodingInterface $encoding;
     private ErrorCorrectionLevelInterface $errorCorrectionLevel;
-    private int $size;
-    private int $margin;
     private RoundBlockSizeModeInterface $roundBlockSizeMode;
     private ColorInterface $foregroundColor;
     private ColorInterface $backgroundColor;
 
     public function __construct(
-        string $data,
-        EncodingInterface $encoding = null,
-        ErrorCorrectionLevelInterface $errorCorrectionLevel = null,
-        int $size = 300,
-        int $margin = 10,
-        RoundBlockSizeModeInterface $roundBlockSizeMode = null,
-        ColorInterface $foregroundColor = null,
-        ColorInterface $backgroundColor = null
+        private string $data,
+        EncodingInterface|null $encoding = null,
+        ErrorCorrectionLevelInterface|null $errorCorrectionLevel = null,
+        private int $size = 300,
+        private int $margin = 10,
+        RoundBlockSizeModeInterface|null $roundBlockSizeMode = null,
+        ColorInterface|null $foregroundColor = null,
+        ColorInterface|null $backgroundColor = null
     ) {
-        $this->data = $data;
         $this->encoding = $encoding ?? new Encoding('UTF-8');
         $this->errorCorrectionLevel = $errorCorrectionLevel ?? new ErrorCorrectionLevelLow();
-        $this->size = $size;
-        $this->margin = $margin;
         $this->roundBlockSizeMode = $roundBlockSizeMode ?? new RoundBlockSizeModeMargin();
         $this->foregroundColor = $foregroundColor ?? new Color(0, 0, 0);
         $this->backgroundColor = $backgroundColor ?? new Color(255, 255, 255);
