@@ -11,29 +11,17 @@ use Endroid\QrCode\QrCodeInterface;
 
 final class DebugResult extends AbstractResult
 {
-    private QrCodeInterface $qrCode;
-    private ?LogoInterface $logo;
-    private ?LabelInterface $label;
-
-    /** @var array<mixed> */
-    private array $options;
-
     private bool $validateResult = false;
 
-    /** @param array<mixed> $options */
     public function __construct(
         MatrixInterface $matrix,
-        QrCodeInterface $qrCode,
-        LogoInterface $logo = null,
-        LabelInterface $label = null,
-        array $options = []
+        private QrCodeInterface $qrCode,
+        private LogoInterface|null $logo = null,
+        private LabelInterface|null $label = null,
+        /** @var array<string, mixed> $options */
+        private array $options = []
     ) {
         parent::__construct($matrix);
-
-        $this->qrCode = $qrCode;
-        $this->logo = $logo;
-        $this->label = $label;
-        $this->options = $options;
     }
 
     public function setValidateResult(bool $validateResult): void
