@@ -8,15 +8,12 @@ use Endroid\QrCode\Matrix\MatrixInterface;
 
 final class EpsResult extends AbstractResult
 {
-    /** @var array<string> */
-    private array $lines;
-
-    /** @param array<string> $lines */
-    public function __construct(MatrixInterface $matrix, array $lines)
-    {
+    public function __construct(
+        MatrixInterface $matrix,
+        /** @var array<string> $lines */
+        private array $lines
+    ) {
         parent::__construct($matrix);
-
-        $this->lines = $lines;
     }
 
     public function getString(): string
